- エンジン名のdefineからOTAFUKUの文字を取り忘れていたところがあったのを修正。   - bookコマンド一部動かなくなってた。ごめん。 - lintで引っかかっていた箇所修正。

+++ source/types.cpp
@@ -24,8 +24,8 @@ std::string PieceToCharBW(" PLNSBRGK        plnsbrgk");
 // operator<<(std::ostream& os,...)とpretty() 
 // ----------------------------------------
 
-std::string pretty(File f) { return pretty_jp ? std::string("１２３４５６７８９").substr((int32_t)f * 2, 2) : std::to_string((int32_t)f + 1); }
-std::string pretty(Rank r) { return pretty_jp ? std::string("一二三四五六七八九").substr((int32_t)r * 2, 2) : std::to_string((int32_t)r + 1); }
+std::string pretty(File f) { return pretty_jp ? std::string("１２３４５６７８９").substr((size_t)f * 2, 2) : std::to_string((size_t)f + 1); }
+std::string pretty(Rank r) { return pretty_jp ? std::string("一二三四五六七八九").substr((size_t)r * 2, 2) : std::to_string((size_t)r + 1); }
 
 std::string pretty(Move m)
 {
